refactor(ios): shrink ScreenshotSeed, Logger, and repository long functions (#349)
Behavior-preserving refactors, no suppressions:

- ScreenshotSeed (682 lines) split per-concern: entry/schedule, session
  writes (+ SessionTiming param struct replacing 6 params), and two verbatim
  markdown fixture files; the 12-case weekly-increment switch became a
  verified lookup table (complexity 13 -> 1). All seed data byte-identical.
- Logger split on its Persistence seam: LogStore class moved verbatim to
  LogStore.swift; writeLog's async body extracted to performWrite.
  SQLiteLogHandler.log keeps its swift-log protocol-fixed 7-param signature
  with #file/#function/#line defaults (SwiftLint discounts defaulted params);
  documented at the declaration.
- ExerciseHistoryRepository: duplicated 16-line row-mapping closures unified
  into historyPoint(from:). SettingsRepository.update: SQL and argument
  building extracted (SQL byte-identical). APIClient: identical
  request-execution tails of perform/performRaw deduplicated into execute(_:)
  (complexity 11 -> ~5).

Baseline regenerated: 79 -> 68 entries, removals only. Unit suite: 914
tests, 0 failures.

Part of #351. Closes #349.
